+++ src/index.ts
@@ -16,6 +16,7 @@ export * from './config';
 export * from './Lookup';
 export * from './createAnnotationFactory';
 export * from './preloadServiceModules';
+export * from './loadServices';
 
 export * from './middlewares/activation';
 export * from './middlewares/config';