+++ test/fixtures/loadServices/Example.ts
@@ -1,0 +1,7 @@
+import {Service} from "@src/.";
+
+@Service('Example')
+//tslint:disable-next-line: class-name
+export class Example {
+
+}