+++ test/fixtures/loadServices/Example2.ts
@@ -1,0 +1,7 @@
+import {Service} from "@src/.";
+
+@Service('Example2')
+//tslint:disable-next-line: class-name
+export class Example2 {
+
+}