+++ test/fixtures/loadServices/JSExample.js
@@ -1,0 +1,8 @@
+const {Service} = require("../../../src");
+const {__decorate} = require('tslib');
+
+class JSExample {
+
+}
+JSExample = __decorate([Service('JSExample')], JSExample);
+exports.JSExample = JSExample;