+++ test/fixtures/loadServices/TSXExample.tsx
@@ -1,0 +1,6 @@
+import {Service} from "@src/decorators/Service";
+
+@Service('TSXExample')
+export class TSXExample {
+
+}