+++ test/fixtures/loadServices/nested/NestedExample.ts
@@ -1,0 +1,6 @@
+import {Service} from "@src/decorators/Service";
+
+@Service('NestedExample')
+export class NestedExample {
+
+}